+++ src.rs/src/lib.rs
@@ -1,7 +1,4 @@
 mod utils;
 
-pub mod commit;
-pub mod element;
-pub mod scalar_field;
 pub mod verkle_ffi_api;
 // pub mod old_api;