+++ src.ts/tests/script.cjs
@@ -1,0 +1,8 @@
+const { loadVerkleCrypto, zeroCommitment } = require('../../dist/cjs/index')
+
+const main = async () => {
+  await loadVerkleCrypto()
+  console.log(zeroCommitment)
+}
+
+main()