+++ src.ts/verkleFFIBindings/index.ts
@@ -1,21 +1,11 @@
-import { Context, zeroCommitment } from '../wasm/rust_verkle_wasm.js'
+import { initVerkleWasm, zeroCommitment } from '../wasm/rust_verkle_wasm.js'
 
-import { getTreeKey, getTreeKeyHash } from './getTreeKey.js'
-
-// This is a 32 byte serialized field element
-type Scalar = Uint8Array
-
-// This is a 64 byte serialized point.
-// It is 64 bytes because the point is serialized in uncompressed format.
-type Commitment = Uint8Array
+import { getTreeKey, getTreeKeyHash, updateCommitment } from './verkleFFI.js'
 
 export {
-  Scalar,
-  Commitment,
-  Context,
-  // This is a function that returns a zero commitment
-  // wasm_bindgen does not seem to allow returning constants
-  zeroCommitment,
+  initVerkleWasm,
   getTreeKey,
   getTreeKeyHash,
+  updateCommitment,
+  zeroCommitment,
 }